requirements.txt
Django==2.0.1
djangorestframework==3.7.7
mysqlclient==1.3.12
python-decouple==3.1
pytz==2017.3
